Replaced ExternalActionTrigger with Events
git-svn-id: svn://localhost:8080/caleydo_main/trunk@2037 1f7349ae-fd9f-0d40-aeb8-9798e6c0fce3

+++ org.caleydo.core/src/org/caleydo/core/data/selection/GenericSelectionManager.java
@@ -278,7 +278,7 @@ public class GenericSelectionManager {
 		Integer[] tempAr = new Integer[elementMap.size()];
 		tempAr = elementMap.keySet().toArray(tempAr);
 		for (Integer element : tempAr) {
-			remove(element, true);
+			remove(element.intValue(), true);
 		}
 	}
